# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API BackEnd",
    "version": "1.0.0",
    "description": "Documentação para a criação de usuários"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Cadastro de usuário",
        "description": "Registra um novo usuário na aplicação",
        "tags": ["Autenticação"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Diogo"
                  },
                  "email": {
                    "type": "string",
                    "example": "diogo@email.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "123456"
                  },
                  "cpf": {
                    "type": "string",
                    "example": "12345678900"
                  }
                },
                "required": ["name", "email", "password", "cpf"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário cadastrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "1"
                        },
                        "name": {
                          "type": "string",
                          "example": "Diogo"
                        },
                        "email": {
                          "type": "string",
                          "example": "diogo@email.com"
                        }
                      }
                    },
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiaWF0IjoxNjg2MzcxMjM2LCJleHBpcmVzSW4iOiIxaaRj"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login de usuário",
        "description": "Realiza o login de um usuário e retorna um token JWT",
        "tags": ["Autenticação"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "diogo@email.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "123456"
                  }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login bem-sucedido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "1"
                        },
                        "name": {
                          "type": "string",
                          "example": "Diogo"
                        },
                        "email": {
                          "type": "string",
                          "example": "diogo@email.com"
                        }
                      }
                    },
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiaWF0IjoxNjg2MzcxMjM2LCJleHBpcmVzSW4iOiIxaaRj"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Busca um usuário por ID",
        "description": "Retorna os detalhes de um usuário específico com base no ID fornecido.",
        "tags": ["Usuários"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do usuário a ser buscado"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "name": { "type": "string" },
                    "email": { "type": "string" },
                    "cpf": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "put": {
        "summary": "Atualiza um usuário",
        "description": "Atualiza os dados de um usuário específico com base no ID fornecido.",
        "tags": ["Usuários"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do usuário a ser atualizado"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Diogo"
                  },
                  "email": {
                    "type": "string",
                    "example": "diogo@email.com"
                  },
                  "cpf": {
                    "type": "string",
                    "example": "12345678900"
                  }
                },
                "required": ["name", "email", "cpf"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "name": { "type": "string" },
                    "email": { "type": "string" },
                    "cpf": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "404": {
            "description": "Usuário não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "delete": {
        "summary": "Soft delete de um usuário",
        "description": "Marca um usuário como deletado sem removê-lo fisicamente do banco de dados.",
        "tags": ["Usuários"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do usuário a ser deletado"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário marcado como deletado com sucesso"
          },
          "404": {
            "description": "Usuário não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
    "/users/{id}/preferences": {
      "get": {
        "summary": "Busca as preferências de um usuário",
        "description": "Retorna as preferências de um usuário específico com base no ID fornecido, incluindo tipo, nome e descrição.",
        "tags": ["Usuários"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do usuário cujas preferências serão buscadas"
          }
        ],
        "responses": {
          "200": {
            "description": "Preferências do usuário encontradas com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "typeId": {
                        "type": "string",
                        "description": "ID do tipo de atividade"
                      },
                      "typeName": {
                        "type": "string",
                        "description": "Nome do tipo de atividade"
                      },
                      "typeDescription": {
                        "type": "string",
                        "description": "Descrição do tipo de atividade"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Preferências não encontradas"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
...
    "/users/avatar": {
      "put": {
        "summary": "Atualiza a foto de perfil do usuário",
        "description": "Permite que o usuário autenticado altere sua foto de perfil.",
        "operationId": "updateUserAvatar",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["Usuários"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": {
                    "type": "string",
                    "format": "binary",
                    "description": "Arquivo de imagem do avatar"
                  }
                },
                "required": ["avatar"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Foto de perfil atualizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Foto de perfil atualizada com sucesso"
                    },
                    "avatar": {
                      "type": "string",
                      "example": "url-ou-base64-da-imagem"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Avatar é obrigatório",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Avatar é obrigatório"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Usuário não encontrado"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Erro interno do servidor"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}